Modelmanagers agora é um pacote em nível de projeto
