+++ scielomanager/scielomanager/utils/modelmanagers/__init__.py
@@ -1,0 +1,2 @@
+from .base import UserObjectQuerySet, UserObjectManager
+
