+++ scielomanager/scielomanager/utils/modelmanagers/helpers.py
@@ -53,10 +53,11 @@ def _patch_userrequestcontextfinder_settings_teardown(func):
     """
     def wrapper(self, **kwargs):
         func(self, **kwargs)
-        
+
         attr_name = settings.USERREQUESTCONTEXT_FINDER.rsplit('.', 1)[-1]
         delattr(finders, attr_name)
 
         settings.USERREQUESTCONTEXT_FINDER = self.default_USERREQUESTCONTEXT_FINDER
 
     return wrapper
+
